:- efficiency fixes for VAULT.createDexLP 	- moved some code base outside of “for loop” (saves gas and compile size) 		- not needed inside loop, they are just redundant calls that get the same value, 	- also diminished excessive function calls and creating local variables instead 	- also changed uint64 _net_usdAmntLP param in function signature to uint256 		- updated ICallitVault.sol accommodate this change 		- NOTE: VAULT.createDexLP invoked from DELEGATE 			(ie. delegate requires new deploy with this vault change)     - NOTE: attempted to refactor function _usd_decimals() into a global,       	however, this results in increased compilation file size (despite being invoked 11 or 12) 			hence, _usd_decimals() not changed, but added commented notes to it 	- NOTE: vault still compiles just fine with no file size or call stack errors
:- NOTE: 
	VAULT and DELEGATE both require re-deploy due to these changes

+++ contracts/ICallitVault.sol
@@ -67,7 +67,7 @@ interface ICallitVault {
     function _usd_decimals() external pure returns (uint8);
     function _payUsdReward(address _sender, uint64 _usdReward, address _receiver) external;
     // function createDexLP(uint256 _resultCnt, uint64 _net_usdAmntLP) external returns(ICallitLib.MARKET_RESULTS memory);
-    function createDexLP(string[] calldata _resultLabels, uint64 _net_usdAmntLP, uint16 _ratioLpTokPerUsd) external returns(ICallitLib.MARKET_RESULTS memory);
+    function createDexLP(string[] calldata _resultLabels, uint256 _net_usdAmntLP, uint16 _ratioLpTokPerUsd) external returns(ICallitLib.MARKET_RESULTS memory);
     // function _createDexLP(address _uswapV2Router, address _uswapv2Factory, address _token, address _usdStable, uint256 _tokenAmount, uint256 _usdAmount) external returns (address);
     function _exePullLiquidityFromLP(address _tokenRouter, address _pairAddress, address _token, address _usdStable) external returns(uint256);
